requirements.txt
# python==3.9.0
beanie==1.11.6
black==19.10b0
boto3==1.24.26
botocore==1.27.26
bson==0.5.10
fastapi==0.78.0
motor==3.0.0
Pillow==9.2.0
pydantic==1.9.1
pymongo==4.1.1
pyparsing==3.0.9
pymongo[srv]==4.1.1
python-dotenv==0.20.0
s3transfer==0.6.0
starlette==0.19.1
urllib3==1.26.7
uvicorn==0.18.2
requests==2.27.1
python-multipart==0.0.5